+++ viprs/__init__.py
@@ -2,5 +2,5 @@
 from .model.VIPRS import VIPRS
 from .utils.data_utils import *
 
-__version__ = '0.1.2'
+__version__ = '0.1.3-alpha'
 __release_date__ = 'December 2024'