+++ viprs/model/gridsearch/HyperparameterSearch.py
@@ -1,8 +1,8 @@
 import pandas as pd
 import numpy as np
-import copy
-from tqdm import tqdm
-import multiprocessing
+from multiprocessing import shared_memory
+from joblib import Parallel, delayed
+
 from pprint import pprint
 
 from viprs.eval.continuous_metrics import r2
@@ -11,56 +11,92 @@ from viprs.model.BayesPRSModel import BayesPRSModel
 from viprs.model.VIPRS import VIPRS
 
 
-def fit_model_fixed_params(params):
-    """
-    Perform model fitting using a set of fixed parameters.
-    This is a helper function to allow us to use the `multiprocessing` module
-    to fit PRS models in parallel.
-    :param params: A tuple of (BayesPRSModel, fixed parameters dictionary, and kwargs for the .fit() method).
+def fit_model_fixed_params(model, fixed_params, shm_data=None, **fit_kwargs):
     """
 
-    # vi_model, fixed_params, **fit_kwargs
-    vi_model, fixed_params, fit_kwargs = params
-    vi_model.fix_params = fixed_params
+    Perform model fitting using a set of fixed set of hyperparameters.
+    This is a helper function to allow users to use the `multiprocessing` module
+    to fit PRS models in parallel.
+
+    :param model: A PRS model object that implements a `.fit()` method and takes `fix_params` as an attribute.
+    :param fixed_params: A dictionary of fixed parameters to use for the model fitting.
+    :param shm_data: A dictionary of shared memory data to use for the model fitting. This is primarily used to
+    share LD data across multiple processes.
+    :param fit_kwargs: Key-word arguments to pass to the `.fit()` method of the PRS model.
+
+    :return: A dictionary containing the coefficient table, hyperparameter table and the training objective.
+    If the model did not converge successfully, return `None`.
+    """
+
+    model.fix_params = fixed_params
+
+    if shm_data is not None:
+
+        model.ld_data = {}
+
+        try:
+            ld_data_shm = shared_memory.SharedMemory(name=shm_data['shm_name'])
+            model.ld_data[shm_data['chromosome']] = np.ndarray(
+                shape=shm_data['shm_shape'],
+                dtype=shm_data['shm_dtype'],
+                buffer=ld_data_shm.buf
+            )
+        except FileNotFoundError:
+            raise Exception("LD data not found in shared memory.")
 
     try:
-        vi_model.fit(**fit_kwargs)
+        model.fit(**fit_kwargs)
     except Exception as e:
-        return None
+        print("Exception encountered when fitting model:", e)
+        model = None
+    finally:
+        if shm_data is not None:
+            ld_data_shm.close()
+            model.ld_data = None
 
-    return vi_model
+    if model is not None:
+        return {
+            'coef_table': model.to_table()[['CHR', 'SNP', 'POS', 'A1', 'A2', 'BETA']],
+            'hyp_table': model.to_theta_table(),
+            'training_objective': model.objective()
+        }
 
 
-class HyperparameterSearch(object):
+class BaseHyperparamSearch(object):
     """
-    A generic class for performing hyperparameter search on the
-    `VIPRS` model. This interface is old and will likely be deprecated
-    in future releases. It is recommended to use the `VIPRSGrid` class
-    and its derivatives for performing grid search instead.
+    A generic class for performing hyperparameter search on any genetic PRS model.
+    This API is under active development and some of the components may change in the near future.
+
+    TODO: Allow users to choose different metrics under each criterion.
+
     """
 
     def __init__(self,
                  gdl,
                  model=None,
-                 criterion='ELBO',
+                 criterion='training_objective',
                  validation_gdl=None,
                  verbose=False,
                  n_jobs=1):
         """
         A generic hyperparameter search class that implements common functionalities
         that may be required by hyperparameter search strategies.
-        :param gdl: A GWADataLoader object
-        :param model: A `PRSModel`-derived object (e.g. VIPRS).
+        :param gdl: A GWADataLoader object containing the GWAS summary statistics for inference.
+        :param model: An instance of the PRS model to use for the hyperparameter search. By default,
+        we use `VIPRS`.
         :param criterion: The objective function for the hyperparameter search.
-        Options are: `ELBO`, `pseudo_validation` or `validation`.
-        :param validation_gdl: If the objective is validation, provide the GWADataLoader object for the validation
-        dataset.
-        :param verbose: Detailed messages and print statements.
+        Options are: `training_objective`, `pseudo_validation` or `validation`. In the case of `VIPRS`, the training
+        objective is the ELBO.
+        :param validation_gdl: If the objective is validation or pseudo-validation, provide the GWADataLoader
+        object for the validation dataset. If the criterion is pseudo-validation, the `validation_gdl` should
+        contain summary statistics from a held-out test set. If the criterion is validation, `validation_gdl` should
+        contain individual-level data from a held-out test set.
+        :param verbose: Verbosity of the information printed to standard output.
         :param n_jobs: The number of processes to use for the hyperparameters search.
         """
 
         # Sanity checking:
-        assert criterion in ('ELBO', 'validation', 'pseudo_validation')
+        assert criterion in ('training_objective', 'validation', 'pseudo_validation')
 
         self.gdl = gdl
         self.n_jobs = n_jobs
@@ -85,8 +121,13 @@ class HyperparameterSearch(object):
         if self._validation_gdl is not None:
             self._validation_gdl.verbose = verbose
 
-        if self.criterion == 'ELBO':
-            assert hasattr(self.model, 'elbo')
+        self._model_coefs = None
+        self._model_hyperparams = None
+        self._training_objective = None
+
+        # Sanity checks:
+        if self.criterion == 'training_objective':
+            assert hasattr(self.model, 'objective')
         elif self.criterion == 'pseudo_validation':
             assert self._validation_gdl is not None
             assert self._validation_gdl.sumstats_table is not None
@@ -98,6 +139,7 @@ class HyperparameterSearch(object):
     def to_validation_table(self):
         """
         Summarize the validation results in a pandas table.
+        :return: A pandas DataFrame with the validation results.
         """
         if self.validation_result is None:
             raise Exception("Validation result is not set!")
@@ -118,36 +160,29 @@ class HyperparameterSearch(object):
         v_df = self.to_validation_table()
         v_df.to_csv(v_filename, index=False, sep=sep)
 
-    def multi_objective(self, models):
+    def _evaluate_models(self):
         """
-        This method evaluates multiple PRS models simultaneously. This can be faster for
-        some evaluation criteria, such as the validation R^2, because we only need to
-        multiply the inferred effect sizes with the genotype matrix only once.
+        This method evaluates multiple PRS models to determine their relative performance based on the
+        criterion set by the user. The criterion can be the training objective (e.g. ELBO in the case of VIPRS),
+        pseudo-validation or validation using held-out test data.
 
-        :param models: A list of PRS models that we wish to evaluate.
+        :return: The metrics associated with each model setup.
         """
 
-        if len(models) == 1:
-            return self.objective(models[0])
+        assert self._training_objective is not None
+        assert self._model_coefs is not None
 
-        if self.criterion == 'ELBO':
-            return [m.elbo() for m in models]
-
+        if self.criterion == 'training_objective':
+            metrics = self._training_objective
         elif self.criterion == 'pseudo_validation':
-            return [m.pseudo_validate(validation_gdl=self._validation_gdl) for m in models]
+            from viprs.eval.pseudo_metrics import pseudo_r2
+
+            metrics = pseudo_r2(self._validation_gdl, self._model_coefs)
+
         else:
 
             prs_m = BayesPRSModel(self._validation_gdl)
-
-            eff_table = models[0].to_table(per_chromosome=False)
-            eff_table = eff_table[['CHR', 'SNP', 'A1', 'A2', 'BETA']]
-            eff_table.rename(columns={'BETA': 'BETA_0'}, inplace=True)
-
-            eff_table[[f'BETA_{i}' for i in range(1, len(models))]] = np.array(
-                [models[i].to_table(per_chromosome=False)['BETA'].values for i in range(1, len(models))]
-            ).T
-
-            prs_m.set_model_parameters(eff_table)
+            prs_m.set_model_parameters(self._model_coefs)
 
             prs = prs_m.predict(test_gdl=self._validation_gdl)
 
@@ -157,155 +192,15 @@ class HyperparameterSearch(object):
                 eval_func = r2
 
             metrics = [eval_func(prs[:, i].flatten(), self._validation_gdl.sample_table.phenotype)
-                       for i in range(len(models))]
+                       for i in range(prs.shape[1])]
 
-            return metrics
-
-    def objective(self, model):
-        """
-        A method that takes the result of fitting the model
-        and returns the desired objective (either `ELBO`, `pseudo_validation`, or `validation`).
-        :param model: The PRS model to evaluate
-        """
-
-        if self.criterion == 'ELBO':
-            return model.elbo()
-        elif self.criterion == 'pseudo_validation':
-            return model.pseudo_validate(validation_gdl=self._validation_gdl)
-        else:
-
-            # Predict:
-            prs = model.predict(test_gdl=self._validation_gdl)
-
-            if self._validation_gdl.phenotype_likelihood == 'binomial':
-                eval_func = roc_auc
-            else:
-                eval_func = r2
-
-            return eval_func(prs, self._validation_gdl.sample_table.phenotype)
+        return metrics
 
     def fit(self):
         raise NotImplementedError
 
 
-class BayesOpt(HyperparameterSearch):
-    """
-    Hyperparameter search using Bayesian optimization
-    """
-
-    def __init__(self,
-                 gdl,
-                 opt_params,
-                 param_bounds=None,
-                 model=None,
-                 criterion='ELBO',
-                 validation_gdl=None,
-                 verbose=False,
-                 n_jobs=1):
-        """
-        Perform hyperparameter search using Bayesian optimization
-        :param gdl: A GWADataLoader object
-        :param opt_params: A list of the hyperparameters to optimize over (e.g. 'pi', 'sigma_epsilon', 'sigma_beta').
-        :param param_bounds: The bounds for each hyperparameter included in the optimization. A list of tuples,
-        where each tuples records the (min, max) values for each hyperparameter.
-        :param model: A `PRSModel`-derived object (e.g. VIPRS).
-        :param criterion: The objective function for the hyperparameter search (ELBO or validation).
-        :param validation_gdl: If the objective is validation, provide the GWADataLoader object for the validation
-        dataset.
-        :param verbose: Detailed messages and print statements.
-        :param n_jobs: The number of processes to use for the hyperparameters search (not applicable here).
-        """
-        
-        super().__init__(gdl,
-                         model=model,
-                         criterion=criterion,
-                         validation_gdl=validation_gdl,
-                         verbose=verbose,
-                         n_jobs=n_jobs)
-
-        self._opt_params = opt_params
-        self._param_bounds = param_bounds or {
-            'sigma_epsilon': (1e-6, 1. - 1e-6),
-            'tau_beta': (1e-3, None),
-            'pi': (1e-6, 1. - 1e-6)
-        }
-
-        # Convert the `pi` limits to log-scale:
-        if 'pi' in self._param_bounds:
-            self._param_bounds['pi'] = tuple(np.log10(self._param_bounds['pi']))
-
-        assert all([opp in self._param_bounds for opp in self._opt_params])
-
-    def fit(self,
-            max_iter=50,
-            f_abs_tol=1e-4,
-            n_calls=30,
-            n_random_starts=5,
-            acq_func="gp_hedge"):
-        """
-        Perform model fitting and hyperparameter search using Bayesian optimization.
-
-        :param n_calls: The number of model runs with different hyperparameter settings.
-        :param n_random_starts: The number of random starts to initialize the optimizer.
-        :param acq_func: The acquisition function (default: `gp_hedge`)
-        :param max_iter: The maximum number of iterations within the search (default: 50).
-        :param f_abs_tol: The absolute tolerance for the objective (ELBO) within the search
-        """
-
-        from skopt import gp_minimize
-
-        def opt_func(p):
-
-            fix_params = dict(zip(self._opt_params, p))
-            if 'pi' in fix_params:
-                fix_params['pi'] = 10**fix_params['pi']
-
-            fitted_model = fit_model_fixed_params((self.model, fix_params,
-                                                   {'max_iter': max_iter,
-                                                    'f_abs_tol': f_abs_tol}))
-
-            if fitted_model is None:
-                return np.inf
-            else:
-                return -self.objective(fitted_model)
-
-        res = gp_minimize(opt_func,  # the function to minimize
-                          [self._param_bounds[op] for op in self._opt_params],  # the bounds on each dimension of x
-                          acq_func=acq_func,  # the acquisition function
-                          n_calls=n_calls,  # the number of evaluations of f
-                          n_random_starts=n_random_starts)  # the random seed
-
-        # Store validation result
-        self.validation_result = []
-        for obj, x in zip(res.func_vals, res.x_iters):
-            v_res = dict(zip(self._opt_params, x))
-            if 'pi' in v_res:
-                v_res['pi'] = 10**v_res['pi']
-
-            if self.criterion == 'ELBO':
-                v_res['ELBO'] = -obj
-            elif self.criterion == 'pseudo_validation':
-                v_res['Pseudo_Validation_Corr'] = -obj
-            else:
-                v_res['Validation_R2'] = -obj
-
-            self.validation_result.append(v_res)
-
-        # Extract the best performing hyperparameters:
-        final_best_params = dict(zip(self._opt_params, res.x))
-        if 'pi' in final_best_params:
-            final_best_params['pi'] = 10 ** final_best_params['pi']
-
-        print("> Bayesian Optimization identified the best hyperparameters as:")
-        pprint(final_best_params)
-
-        print("> Refitting the model with the best hyperparameters...")
-
-        self.model.fix_params = final_best_params
-        return self.model.fit()
-
-
-class GridSearch(HyperparameterSearch):
+class GridSearch(BaseHyperparamSearch):
     """
     Hyperparameter search using Grid Search
     """
@@ -314,24 +209,30 @@ class GridSearch(HyperparameterSearch):
                  gdl,
                  grid,
                  model=None,
-                 criterion='ELBO',
+                 criterion='training_objective',
                  validation_gdl=None,
                  verbose=False,
                  n_jobs=1):
 
         """
         Perform hyperparameter search using grid search
-        :param gdl: A GWADataLoader object
-        :param grid: A HyperParameterGrid object
-        :param model: A `PRSModel`-derived object (e.g. VIPRS).
-        :param criterion: The objective function for the grid search (ELBO or validation).
-        :param validation_gdl: If the objective is validation, provide the GWADataLoader object for the validation
-        dataset.
-        :param verbose: Detailed messages and print statements.
-        :param n_jobs: The number of processes to use for the grid search
+        :param gdl: A GWADataLoader object containing the GWAS summary statistics for inference.
+        :param model: An instance of the PRS model to use for the hyperparameter search. By default,
+        we use `VIPRS`.
+        :param criterion: The objective function for the hyperparameter search.
+        Options are: `training_objective`, `pseudo_validation` or `validation`. In the case of `VIPRS`, the training
+        objective is the ELBO.
+        :param validation_gdl: If the objective is validation or pseudo-validation, provide the GWADataLoader
+        object for the validation dataset. If the criterion is pseudo-validation, the `validation_gdl` should
+        contain summary statistics from a held-out test set. If the criterion is validation, `validation_gdl` should
+        contain individual-level data from a held-out test set.
+        :param verbose: Verbosity of the information printed to standard output.
+        :param n_jobs: The number of processes to use for the hyperparameters search.
         """
 
-        super().__init__(gdl, model=model, criterion=criterion,
+        super().__init__(gdl,
+                         model=model,
+                         criterion=criterion,
                          validation_gdl=validation_gdl,
                          verbose=verbose,
                          n_jobs=n_jobs)
@@ -339,171 +240,121 @@ class GridSearch(HyperparameterSearch):
         self.grid = grid
         self.model.threads = 1
 
-    def fit(self, max_iter=50, f_abs_tol=1e-3, x_abs_tol=1e-8):
+    def fit(self, max_iter=1000, f_abs_tol=1e-6, x_abs_tol=1e-6):
+        """
+        Perform grid search over the hyperparameters to determine the
+        best model based on the criterion set by the user. This utility method
+        performs model fitting across the grid of hyperparameters, potentially in parallel
+        if `n_jobs` is greater than 1.
+
+        :param max_iter: The maximum number of iterations to run for each model fit.
+        :param f_abs_tol: The absolute tolerance for the function convergence criterion.
+        :param x_abs_tol: The absolute tolerance for the parameter convergence criterion.
+
+        :return: The best model based on the criterion set by the user.
+        """
 
         print("> Performing Grid Search over the following grid:")
         print(self.grid.to_table())
 
-        opts = [(self.model, g, {'max_iter': max_iter,
-                                 'f_abs_tol': f_abs_tol,
-                                 'x_abs_tol': x_abs_tol})
-                for g in self.grid.combine_grids()]
+        if self.n_jobs > 1:
+            # Only create the shared memory object if the number of processes is more than 1.
+            # Otherwise, this would be a waste of resources.
 
-        assert len(opts) > 1
+            # ----------------- Copy the LD data to shared memory -----------------
+            ld_data_arr = self.model.ld_data[self.model.chromosomes[0]]
+            # Create a shared memory block for the array
+            shm = shared_memory.SharedMemory(create=True, size=ld_data_arr.nbytes)
+
+            # Create a NumPy array backed by the shared memory block
+            shared_array = np.ndarray(ld_data_arr.shape, dtype=ld_data_arr.dtype, buffer=shm.buf)
+
+            np.copyto(shared_array, ld_data_arr)
+
+            del ld_data_arr
+            self.model.ld_data = None
+
+            shm_args = {
+                'shm_name': shm.name,
+                'chromosome': self.model.chromosomes[0],
+                'shm_shape': shared_array.shape,
+                'shm_dtype': shared_array.dtype
+            }
+
+        else:
+            shm_args = None
+
+        # --------------------------------------------------------------------
+        # Perform grid search:
+
+        grid = self.grid.combine_grids()
+
+        parallel = Parallel(n_jobs=self.n_jobs, backend='multiprocessing')
+
+        with parallel:
+
+            fitted_models = parallel(
+                delayed(fit_model_fixed_params)(self.model, g, shm_args,
+                                                max_iter=max_iter,
+                                                f_abs_tol=f_abs_tol,
+                                                x_abs_tol=x_abs_tol)
+                for g in grid
+            )
+
+        # Clean up after performing model fit:
+        self.model.ld_data = None  # To minimize memory usage with validation/pseudo-validation
+
+        # Close and unlink shared memory objects:
+        if shm_args is not None:
+            shm.close()
+            shm.unlink()
+
+        # --------------------------------------------------------------------
+        # Post-process the results and determine the best model:
+
+        assert not all([fm is None for fm in fitted_models]), "None of the models converged successfully."
+
+        # 1) Extract the data from the trained models:
+        from viprs.utils.compute_utils import combine_coefficient_tables
+
+        self._model_coefs = combine_coefficient_tables([fm['coef_table'] for fm in fitted_models if fm is not None])
+        self._model_hyperparams = [fm['hyp_table'] for fm in fitted_models if fm is not None]
+        self._training_objective = [fm['training_objective'] for fm in fitted_models if fm is not None]
+
+        # 2) Perform evaluation on the models that converged:
+        eval_metrics = self._evaluate_models()
+
+        # 3) Combine all the results together into a single table (populate records in
+        # self.validation_result):
 
         self.validation_result = []
-        fit_results = []
-        params = []
+        success_counter = 0
 
-        ctx = multiprocessing.get_context("spawn")
+        for i, vr in enumerate(grid):
+            if fitted_models[i] is not None:
+                vr['Converged'] = True
+                vr['training_objective'] = self._training_objective[success_counter]
+                if self.criterion != 'training_objective':
+                    vr[self.criterion] = eval_metrics[success_counter]
+                success_counter += 1
+            else:
+                vr['Converged'] = False
+                vr['training_objective'] = np.NaN
+                if self.criterion != 'training_objective':
+                    vr[self.criterion] = np.NaN
 
-        with ctx.Pool(self.n_jobs, maxtasksperchild=1) as pool:
+            self.validation_result.append(vr)
 
-            for idx, fitted_model in tqdm(enumerate(pool.imap(fit_model_fixed_params, opts)), total=len(opts)):
+        # --------------------------------------------------------------------
+        # Determine and return the best model:
 
-                if fitted_model is None:
-                    continue
-
-                fit_results.append(fitted_model)
-                params.append(copy.copy(opts[idx][1]))
-                self.validation_result.append(copy.copy(opts[idx][1]))
-                self.validation_result[-1]['ELBO'] = fitted_model.elbo()
-
-        if len(fit_results) > 1:
-            res_objectives = self.multi_objective(fit_results)
-        else:
-            raise Exception("Error: Convergence was achieved for less than 2 models.")
-
-        if self.criterion == 'validation':
-            for i in range(len(self.validation_result)):
-                self.validation_result[i]['Validation_R2'] = res_objectives[i]
-        elif self.criterion == 'pseudo_validation':
-            for i in range(len(self.validation_result)):
-                self.validation_result[i]['Pseudo_Validation_Corr'] = res_objectives[i]
-
-        best_idx = np.argmax(res_objectives)
-        best_params = params[best_idx]
+        best_idx = np.argmax(self.to_validation_table()[self.criterion].values)
 
         print("> Grid search identified the best hyperparameters as:")
-        pprint(best_params)
+        pprint(grid[best_idx])
 
-        print("> Refitting the model with the best hyperparameters...")
+        self.model.fix_params = grid[best_idx]
+        self.model.initialize()
+        self.model.set_model_parameters(fitted_models[best_idx]['coef_table'])
 
-        self.model.fix_params = best_params
-        return self.model.fit()
-
-
-class BMA(BayesPRSModel):
-    """
-    Bayesian Model Averaging fitting procedure
-    """
-
-    def __init__(self,
-                 gdl,
-                 grid,
-                 model=None,
-                 normalization='softmax',
-                 verbose=False,
-                 n_jobs=1):
-        """
-        Integrate out hyperparameters using Bayesian Model Averaging
-        :param gdl: A GWADataLoader object
-        :param grid: A HyperParameterGrid object
-        :param model: A `PRSModel`-derived object (e.g. VIPRS).
-        :param normalization: The normalization scheme for the final ELBOs. Options are (`softmax`, `sum`).
-        :param verbose: Detailed messages and print statements.
-        :param n_jobs: The number of processes to use for the BMA
-        """
-
-        super().__init__(gdl)
-
-        assert normalization in ('softmax', 'sum')
-
-        self.grid = grid
-        self.n_jobs = n_jobs
-        self.verbose = verbose
-
-        if model is None:
-            self.model = VIPRS(gdl)
-        else:
-            self.model = model
-
-        self.model.verbose = verbose
-        self.model.threads = 1
-
-        self.normalization = normalization
-
-        self.var_gamma = None
-        self.var_mu = None
-        self.var_sigma = None
-
-    def initialize(self):
-
-        self.var_gamma = {c: np.zeros(c_size) for c, c_size in self.shapes.items()}
-        self.var_mu = {c: np.zeros(c_size) for c, c_size in self.shapes.items()}
-        self.var_sigma = {c: np.zeros(c_size) for c, c_size in self.shapes.items()}
-
-    def fit(self, max_iter=100, f_abs_tol=1e-3, x_abs_tol=1e-8, **grid_kwargs):
-
-        self.initialize()
-
-        print("> Performing Bayesian Model Averaging with the following grid:")
-        print(self.grid.to_table())
-
-        opts = [(self.model, g, {'max_iter': max_iter,
-                                 'f_abs_tol': f_abs_tol,
-                                 'x_abs_tol': x_abs_tol})
-                for g in self.grid.combine_grids()]
-
-        elbos = []
-        var_gammas = []
-        var_mus = []
-        var_sigmas = []
-
-        ctx = multiprocessing.get_context("spawn")
-
-        with ctx.Pool(self.n_jobs, maxtasksperchild=1) as pool:
-            for fitted_model in tqdm(pool.imap_unordered(fit_model_fixed_params, opts), total=len(opts)):
-
-                if fitted_model is None:
-                    continue
-
-                elbos.append(fitted_model.elbo())
-                var_gammas.append(fitted_model.var_gamma)
-                var_mus.append(fitted_model.var_mu)
-                var_sigmas.append(fitted_model.var_sigma)
-
-        elbos = np.array(elbos)
-
-        if self.normalization == 'softmax':
-            from scipy.special import softmax
-            elbos = softmax(elbos)
-        elif self.normalization == 'sum':
-            # Correction for negative ELBOs:
-            elbos = elbos - elbos.min() + 1.
-            elbos /= elbos.sum()
-
-        for idx in range(len(elbos)):
-            for c in self.shapes:
-                self.var_gamma[c] += var_gammas[idx][c]*elbos[idx]
-                self.var_mu[c] += var_mus[idx][c]*elbos[idx]
-                self.var_sigma[c] += var_sigmas[idx][c]*elbos[idx]
-
-        self.pip = {}
-        self.post_mean_beta = {}
-        self.post_var_beta = {}
-
-        for c, v_gamma in self.var_gamma.items():
-
-            if len(v_gamma.shape) > 1:
-                self.pip[c] = v_gamma.sum(axis=1)
-                self.post_mean_beta[c] = (v_gamma*self.var_mu[c]).sum(axis=1)
-                self.post_var_beta[c] = ((v_gamma * (self.var_mu[c] ** 2 + self.var_sigma[c])).sum(axis=1) -
-                                         self.post_mean_beta[c]**2)
-            else:
-                self.pip[c] = v_gamma
-                self.post_mean_beta[c] = v_gamma * self.var_mu[c]
-                self.post_var_beta[c] = (v_gamma * (self.var_mu[c] ** 2 + self.var_sigma[c]) -
-                                         self.post_mean_beta[c]**2)
-
-        return self
+        return self.model