+++ CHANGELOG.md
@@ -5,7 +5,27 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.1.2] - 2024-06-03
+## [0.1.3] - 2025-01-16
+
+### Changed
+
+- Fixed bugs in `VIPRSGridSearch` and `VIPRSBMA` models, specifically how they were handling `_log_var_tau`, 
+and the hyperparameters objects after selecting best models or performing model averaging.
+- Fixed bug in how `viprs_fit` handles validation `gdl`s when the user passes genotype data.
+- Updated interfaces in `HyperparameterSearch` script to make it more flexible and efficient. Primarily, 
+I added shared memory object for the LD matrix to avoid redundant memory usage when fitting multiple
+models in parallel. (** WORK IN PROGRESS **).
+
+### Added
+
+- Added `viprs-cli-example.ipynb` notebook to demonstrate how to use the `viprs` commandline interface.
+- Added documentation page for Downloading LD matrices.
+- Added new utility function `combine_coefficient_tables` to combine the output from multiple VIPRS models.
+- Added more thorough tests for `VIPRSGridSearch` and `VIPRSBMA` models.
+- Added `PeakMemoryProfiler` to `viprs_fit` to more accurately track peak memory usage. Temporary solution, 
+this will be moved to `magenpy` later on.
+
+## [0.1.2] - 2024-12-25
 
 ### Changed